[0.2.x] fixed: samples
git-svn-id: https://pykd.svn.codeplex.com/svn@82297 9b283d60-5439-405e-af05-b73fd8c4d996

+++ samples/dbg/widgets/registers.py
@@ -14,12 +14,13 @@ class RegisterWidget( DebugWidget ):
     def updateView(self):
 
         s = ""
-
+        
         try:
             i = 0
             while True:
                 reg = pykd.reg(i)
-                s += "%s    %x ( %d )\r\n" % ( reg.name(), reg, reg )
+                #print str(reg)
+                s += "%s    %x ( %d )\r\n" % ( reg.name(), long(reg), long(reg) )
                 i += 1
 
         except pykd.BaseException:
